requirements.txt
Flask==3.1.0
openai==1.58.1
python-dotenv==1.0.1
langchain-chroma==0.1.4
langchain-text-splitters==0.3.4
langchain-community==0.3.8
langchain-openai==0.2.10
chromadb==0.5.23
pypdf==5.5.0